Fix warnings in the assembler
This removes all warnings when compiling the assembler on both 32bit and
64bit.

- Change many types that store a size to size_t so that they match the
  size being stored in them.

- Explicitly cast to DWORDs when referring to what I assume forms part
  of the bytecode (warning is valid - there is an endian bug here, but
  we don't need to care about that since we are only on x86).

- Explicitly cast sizes to ints when storing in a shader_ins since I
  presume this forms part of the bytecode and size_t would not be
  appropriate. I assume that sizes are already guaranteed to be small
  enough to fit in 7 bits.

- Explicitly cast the size to a DWORD when calling the MD5-like hash
  function since this function is expecting all variables to look like a
  32bit x86 CPU's registers (and if a shader was larger than 4GB we
  would have other problems ;-)

- Explicitly cast doubles to ints for the progress displays in
  Shaders.cpp. This is what the compiler will do anyway so I'm assuming
  that 32bit signed integers are ok and the doubles were only required
  for intermediate calculations that exceeded 31 bits of precision.

Spot tested in Witcher 3

+++ D3D_Shaders/Shaders.cpp
@@ -48,7 +48,7 @@ int _tmain(int argc, _TCHAR* argv[])
 	vector<string> lines;
 	fopen_s(&f, "gamelist.txt", "rb");
 	if (f) {
-		int fr = ::fread(gamebuffer, 1, 10000, f);
+		size_t fr = ::fread(gamebuffer, 1, 10000, f);
 		fclose(f);
 		lines = stringToLines(gamebuffer, fr);
 	}
@@ -83,7 +83,7 @@ int _tmain(int argc, _TCHAR* argv[])
 				fwrite(ASM.data(), 1, ASM.size(), f);
 				fclose(f);
 				
-				int newProgress = 50.0 * i / files.size();
+				int newProgress = (int)(50.0 * i / files.size());
 				if (newProgress > progress) {
 					cout << ".";
 					progress++;
@@ -115,7 +115,7 @@ int _tmain(int argc, _TCHAR* argv[])
 				fwrite(CBO.data(), 1, CBO.size(), f);
 				fclose(f);
 
-				int newProgress = 50.0 * i / files.size();
+				int newProgress = (int)(50.0 * i / files.size());
 				if (newProgress > progress) {
 					cout << ".";
 					progress++;
@@ -143,7 +143,7 @@ int _tmain(int argc, _TCHAR* argv[])
 				fwrite(ASM.data(), 1, ASM.size(), f);
 				fclose(f);
 
-				int newProgress = 50.0 * i / files.size();
+				int newProgress = (int)(50.0 * i / files.size());
 				if (newProgress > progress) {
 					cout << ".";
 					progress++;
