+++ D3D_Shaders/stdafx.h
@@ -40,7 +40,7 @@ struct token_operand
 vector<DWORD> assembleIns(string s);
 vector<byte> readFile(string fileName);
 vector<DWORD> ComputeHash(byte const* input, DWORD size);
-vector<string> stringToLines(const char* start, int size);
+vector<string> stringToLines(const char* start, size_t size);
 DWORD strToDWORD(string s);
 string convertF(DWORD original);
 HRESULT disassembler(vector<byte> *buffer, vector<byte> *ret, const char *comment);
